refactor the connection_manager code.
