+++ main/xiaozhi-server/webrtc/modules/__init__.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+"""
+WebRTC模块组件包
+"""
+
+from .audio_track_processor import AudioTrackProcessor
+from .asr_helper import ASRHelper
+from .vad_helper import VADHelper
+from .webrtc_connection import WebRTCConnection
+
+__all__ = ['AudioTrackProcessor', 'ASRHelper', 'VADHelper', 'WebRTCConnection']
